refactor: share code into libraries
